add an outdoor sensor, fix measuring continuation errors

+++ Temperature Measurement Station/TMC/InputOutput.py
@@ -19,6 +19,7 @@ dht_signifier = "DHT22: "
 name_signifier = "Name: "
 light_signifier = "Light: "
 track_signifier = "Iteration: "
+temp_out_signifier = "DS18: "
 
 class I_O:
 
@@ -65,6 +66,18 @@ class I_O:
                     print ("There is no led pin here.")
                     break
     
+    def temp_out_address_read(self):        
+        with self.pin_file.open() as file:
+            while (True):
+                read = file.readline()
+                if read.startswith(temp_out_signifier):
+                    read = read.replace(temp_out_signifier, "")
+                    addresses = read.split(" ")
+                    return addresses
+                elif read is "":
+                    print ("There are no addresses here.")
+                    break
+    
     def name_txt_read(self):        
         with self.name_file.open() as file:
             while (True):
@@ -93,7 +106,7 @@ class I_O:
                     self.iteration = iteration + 1
                     break
                 elif read is "":
-                    print ("There is no track number here here.")
+                    print ("There is no track number here.")
                     break
         log_file_name = "{}_{:03}_log.txt".format(self.title, self.iteration)
         self.log_file = self.log_dir / log_file_name
@@ -104,37 +117,43 @@ class I_O:
     def track_txt_prime(self):
         track_file_name = "{}_track.txt".format(self.title)
         self.track_file = info_path / track_file_name
-        track_file_content = "{}{}".format(track_signifier, "0")
-        with self.track_file.open("w") as file:
-            file.write(track_file_content.decode('utf-8'))
+        if not self.track_file.exists():
+            track_file_content = "{}{}".format(track_signifier, "0")
+            with self.track_file.open("w") as file:
+                file.write(track_file_content.decode('utf-8'))
 
-    def log_txt_prime(self, dht_names=[[]], light_names=[]):
+    def log_txt_prime(self, dht_names=[[]], temp_out_names=[], light_names=[]):
         string = []
         string.append("{}".format("Time"))
         for name in dht_names:
             string.append("{}".format(name[0]))
         for name in dht_names:
             string.append("{}".format(name[1]))
-        for name in light_names:
+        for name in temp_out_names:
             string.append("{}".format(name))
-        string.append("\n")
-        self.title_string = ', '.join(string)
-        print(self.title_string)
-        with self.log_file.open("a") as file:
-            file.write(self.title_string.decode('utf-8'))
+        for name in light_names:
+            string.append("{}".format(name))   
+        self.title_string = ', '.join(string) + "\n"
+        if not self.log_file.exists():
+            print(self.title_string)
+            with self.log_file.open("a") as file:
+                file.write(self.title_string.decode('utf-8'))
+        else:
+            print("Continuation...\n{}".format(self.title_string))
 
-    def log_txt_write(self, temp_hum_values=[[]], light_values=[]):
+    def log_txt_write(self, temp_hum_values=[[]], temp_out_values=[], light_values=[]):
         string = []
         dateTime = dt.datetime.now().strftime("%x") + " " + dt.datetime.now().strftime("%X")
         string.append(dateTime)
-        for value in temp_hum_values:
-            string.append("{}".format(value[0]))
-        for value in temp_hum_values:
-            string.append("{}".format(value[1] / 100))
-        for value in light_values:
-            string.append("{}".format(value))
-        string.append("\n")
-        self.log_string = ', '.join(string)
+        for name in temp_hum_values:
+            string.append("{}".format(name[0]))
+        for name in temp_hum_values:
+            string.append("{}".format(name[1]))
+        for name in temp_out_values:
+            string.append("{}".format(name))
+        for name in light_values:
+            string.append("{}".format(name))   
+        self.log_string = ', '.join(string) + "\n"
         print(self.log_string)
         with self.log_file.open("a") as file:
             file.write(self.log_string.decode('utf-8'))
